LayoutTests/requestidlecallback/requestidlecallback-is-not-called-when-canceled-expected.txt

This tests that when requestIdleCallback is enabled, requestIdleCallback is eventaully called in the order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a3 is a2 + 1
PASS logsA.length is 2
PASS logsB.length is 1
PASS logsA.join(", ") is "A1, A3"
PASS logsB.join(", ") is "B1"
PASS successfullyParsed is true

TEST COMPLETE
